move place stuff into lib folder
