+++ src/lib/places/Compass.svelte
